# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Augmented Things",
        "description": "Augmented Things it's a project developed by @Jatapiaro",
        "contact": {
            "email": "jatapiaro@gmail.com"
        },
        "version": "1"
    },
    "paths": {
...
        "/api/v1/types": {
            "get": {
                "tags": [
                    "Types"
                ],
                "summary": "Shows the system types",
                "responses": {
                    "200": {
                        "description": "Shows the system types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Types"
                ],
                "summary": "Register a new device type",
                "requestBody": {
                    "description": "Type that needs to be stored",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "object",
                                        "$ref": "#/components/schemas/Type"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Device that was registered.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/api/v1/types/{type}": {
            "get": {
                "tags": [
                    "Types"
                ],
                "summary": "Shows the specified type",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "ID of the type",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abcdefghijklmnopq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shows the specified type",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Types"
                ],
                "summary": "Updates a type",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Type to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Type to be updated",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "object",
                                        "$ref": "#/components/schemas/Type"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Type that was updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Types"
                ],
                "summary": "Deletes a type",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Type to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Type that was deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Type": {
                "title": "Type",
                "description": "Types of valid devices and his types on the system",
                "required": [
                    "id",
                    "type"
                ],
                "properties": {
                    "id": {
                        "schema": "Type",
                        "description": "ID of the particle device",
                        "type": "string"
                    },
                    "type": {
                        "schema": "Type",
                        "description": "Type of device",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}